requirements.txt
Django>=5.0,<6.0
wagtail>=6.0,<7.0
django-taggit>=5.0
django-modelcluster>=6.0
Pillow>=10.0

gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
psycopg[binary]>=3.1
django-environ>=0.11
